Initial Learnings
Remembered this album (Sonic Highways) from when I was a kid, I remember 'not liking it' for some reason, and I can't figure out exactly why. Now I absolutely love it, looking back on it. I hear a ton of influences in the style, and I love the concept for the album. I think it might have been because I didn't 'understand' music like I do now, it was just before I had started playing guitar I think too.
https://www.youtube.com/watch?v=XCaRRvBCPPE

Wow this is a very nostalgic MCPE video, now that makes me want to set up some classic MC versions so I can play them again. I do think I really like the simpler/older versions of MC, I think the game's direction is changing into something else, it's just not what I have known it to be all of these years. Not to say that's good or bad, just that it's different, and the things I used to like/relate to it about aren't as much at the forefront anymore. I do think change is good for it though, because that's likely what's help keep it mainstream in the game world for so long. It is healthy for it, I'd rather it change and evolve than to just fall off into obscurity.
https://www.youtube.com/watch?v=vDHuO-OAlf8

I was watching part of this old one the other day too, classic times :)
https://www.youtube.com/watch?v=K4lFZcyiA7M

Made this new song (Scopes) the other day, this is what I'm listening to while writing this. It goes in multiple directions for sure.
https://www.youtube.com/watch?v=p7sEcINJgeM&list=PL0O7adIvyakQfyvsTUe8TwYa9QgJ7CTlv&index=4

I'm going to work on posting the two songs I made last night too, happy with how those turned out.